refactor: extract EndTurnOutcome into separate data class
